Tweaked the fast-avro schema fingerprinting logic (#508)
It used to leverage the Avro fingerprinting, which is based on the
"parsing canonical form" of the schema, but this ignores properties
such as which class to use for String deserialization. This is a
problem in cases where we wish to deserialize the same schema both
with and without Java Strings, since the cache will return the same
deserializer for both, ultimately leading to class cast issues.

The new approach is to simply take the MD5 hash of the full string
of the schema.

+++ fastserde/avro-fastserde/src/main/java/com/linkedin/avro/fastserde/FastDeserializerGeneratorBase.java
@@ -37,8 +37,8 @@ public abstract class FastDeserializerGeneratorBase<T> extends FastSerdeBase {
   }
 
   public static String getClassName(Schema writerSchema, Schema readerSchema, String description) {
-    Long writerSchemaId = Math.abs(Utils.getSchemaFingerprint(writerSchema));
-    Long readerSchemaId = Math.abs(Utils.getSchemaFingerprint(readerSchema));
+    int writerSchemaId = Math.abs(Utils.getSchemaFingerprint(writerSchema));
+    int readerSchemaId = Math.abs(Utils.getSchemaFingerprint(readerSchema));
     String typeName = SchemaAssistant.getTypeName(readerSchema);
     return typeName + SEP + description + "Deserializer" + SEP + writerSchemaId + SEP + readerSchemaId;
   }
